+++ uspto/util/common.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 # (c) 2017 Andreas Motl <andreas@ip-tools.org>
+import os
 import sys
 import logging
+import pathvalidate
 
 def to_list(obj):
     """Convert an object to a list if it is not already one"""
@@ -22,3 +24,10 @@ def setup_logging(level=logging.INFO):
         format=log_format,
         stream=sys.stderr,
         level=level)
+
+def get_document_path(directory, number, format, source=None):
+    if source:
+        source = source.lower() + '.'
+    filename = pathvalidate.sanitize_filename('{name}.{source}{suffix}'.format(name=number.upper(), source=source, suffix=format.lower()))
+    filepath = os.path.join(directory, filename)
+    return filepath