+++ uspto/peds/command.py
@@ -31,8 +31,14 @@ def run():
         # Download granted patent by patent number
         {program} get "PP28532" --type=patent --format=xml
 
-        # Download granted patent by patent number and save to /var/spool/uspto-pair/PP28532.xml
-        {program} save "PP28532" --type=patent --format=xml --directory=/var/spool/uspto-pair
+        # Download granted patent by patent number and save to /var/spool/uspto/PP28532.peds.xml
+        {program} save "PP28532" --type=patent --format=xml --directory=/var/spool/uspto
+
+
+    Bulk examples:
+
+        # Download all documents from numbers.txt and save them /var/spool/uspto/$number.peds.(xml|json)
+        {program} bulk save --numberfile=numbers.txt --format=xml,json --pretty --directory=/var/spool/uspto --wait
 
     """