+++ CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - Improve Celery bootstrapping
 - Add “--poll” option to wait for results when running in “--background” mode
+- Improve command line argument handling
 
 2017-10-21 0.3.0
 ================